// my-app/src/components/Recipe/Recipe.d.ts
interface Ingredient {
    uuid: string;
    amount: number;
    measurement: string;
    name: string;
}

interface Direction {
    instructions: string;
    optional: boolean;
}

interface Images {
    full: string;
    medium: string;
    small: string;
}

export interface RecipeProps {
    uuid: string;
    title: string;
    description: string;
    images: Images;
    servings: number;
    prepTime: number;
    cookTime: number;
    postDate: Date;
    editDate: Date;
    ingredients: [
        Ingredient
    ];
    directions: [
        Direction
    ];
}

export interface SpecialsProps {
    uuid: string;
    ingredientId: string;
    type: string;
    title: string;
    geo?: string;
    text?: string;
}

export interface IProps extends RecipeProps {
    matchingSpecials: SpecialsProps[];
}